Remove Float64 restriction
Throughout the package types are often hard-wired to be Float64. The PR lifts this restriction in some places to allow for any real type. In particular, tihs allows the use of Float32 or Float16 to reduce memory footprint. Structures like `QuadLoss` etc. still require modification in order to not cause conversions to Float64 when computations are made using its fields, but this PR at least allows the user to have data in other formats.